requirements.txt
argparse==1.4.0
colorama
plotly
pysam
numpy
suffix-trees
msbwt
